Split out `strategies` module
